+++ src/Classifiers/SoftmaxClassifier.php
@@ -70,7 +70,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
      *
      * @var float
      */
-    protected float $alpha;
+    protected float $l2Penalty;
 
     /**
      * The maximum number of training epochs. i.e. the number of times to iterate
@@ -125,7 +125,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
     /**
      * @param int $batchSize
      * @param \Rubix\ML\NeuralNet\Optimizers\Optimizer|null $optimizer
-     * @param float $alpha
+     * @param float $l2Penalty
      * @param int $epochs
      * @param float $minChange
      * @param int $window
@@ -135,7 +135,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
     public function __construct(
         int $batchSize = 128,
         ?Optimizer $optimizer = null,
-        float $alpha = 1e-4,
+        float $l2Penalty = 1e-4,
         int $epochs = 1000,
         float $minChange = 1e-4,
         int $window = 5,
@@ -146,9 +146,9 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
                 . " greater than 0, $batchSize given.");
         }
 
-        if ($alpha < 0.0) {
-            throw new InvalidArgumentException('Alpha must be'
-                . " greater than 0, $alpha given.");
+        if ($l2Penalty < 0.0) {
+            throw new InvalidArgumentException('L2 Penalty must be'
+                . " greater than 0, $l2Penalty given.");
         }
 
         if ($epochs < 1) {
@@ -168,7 +168,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
 
         $this->batchSize = $batchSize;
         $this->optimizer = $optimizer ?? new Adam();
-        $this->alpha = $alpha;
+        $this->l2Penalty = $l2Penalty;
         $this->epochs = $epochs;
         $this->minChange = $minChange;
         $this->window = $window;
@@ -213,7 +213,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
         return [
             'batch size' => $this->batchSize,
             'optimizer' => $this->optimizer,
-            'alpha' => $this->alpha,
+            'l2 penalty' => $this->l2Penalty,
             'epochs' => $this->epochs,
             'min change' => $this->minChange,
             'window' => $this->window,
@@ -287,7 +287,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
 
         $this->network = new FeedForward(
             new Placeholder1D($dataset->numFeatures()),
-            [new Dense(count($classes), $this->alpha, true, new Xavier1())],
+            [new Dense(count($classes), $this->l2Penalty, true, new Xavier1())],
             new Multiclass($classes, $this->costFn),
             $this->optimizer
         );