+++ src/Regressors/MLPRegressor.php
@@ -89,7 +89,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
      *
      * @var float
      */
-    protected float $alpha;
+    protected float $l2Penalty;
 
     /**
      * The maximum number of training epochs. i.e. the number of times to iterate
@@ -159,7 +159,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
      * @param \Rubix\ML\NeuralNet\Layers\Hidden[] $hiddenLayers
      * @param int $batchSize
      * @param \Rubix\ML\NeuralNet\Optimizers\Optimizer|null $optimizer
-     * @param float $alpha
+     * @param float $l2Penalty
      * @param int $epochs
      * @param float $minChange
      * @param int $window
@@ -172,7 +172,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
         array $hiddenLayers = [],
         int $batchSize = 128,
         ?Optimizer $optimizer = null,
-        float $alpha = 1e-4,
+        float $l2Penalty = 1e-4,
         int $epochs = 1000,
         float $minChange = 1e-4,
         int $window = 5,
@@ -192,9 +192,9 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
                 . " greater than 0, $batchSize given.");
         }
 
-        if ($alpha < 0.0) {
-            throw new InvalidArgumentException('Alpha must be'
-                . " greater than 0, $alpha given.");
+        if ($l2Penalty < 0.0) {
+            throw new InvalidArgumentException('L2 Penalty must be'
+                . " greater than 0, $l2Penalty given.");
         }
 
         if ($epochs < 1) {
@@ -224,7 +224,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
         $this->hiddenLayers = $hiddenLayers;
         $this->batchSize = $batchSize;
         $this->optimizer = $optimizer ?? new Adam();
-        $this->alpha = $alpha;
+        $this->l2Penalty = $l2Penalty;
         $this->epochs = $epochs;
         $this->minChange = $minChange;
         $this->window = $window;
@@ -272,7 +272,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
             'hidden layers' => $this->hiddenLayers,
             'batch size' => $this->batchSize,
             'optimizer' => $this->optimizer,
-            'alpha' => $this->alpha,
+            'l2 penalty' => $this->l2Penalty,
             'epochs' => $this->epochs,
             'min change' => $this->minChange,
             'window' => $this->window,
@@ -353,7 +353,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
 
         $hiddenLayers = $this->hiddenLayers;
 
-        $hiddenLayers[] = new Dense(1, $this->alpha, true, new Xavier2());
+        $hiddenLayers[] = new Dense(1, $this->l2Penalty, true, new Xavier2());
 
         $this->network = new FeedForward(
             new Placeholder1D($dataset->numFeatures()),